--- src/lib/services/accounts/service/create-practice/index.ts
@@ -1,1 +1,0 @@
-export * as CreatePractice from './createPractice';
