+++ src/lib/services/accounts/service/handle-practice-onboarding/handlePracticeOnboarding.ts
@@ -1,22 +1,25 @@
-import { CreatePractice } from '@/lib/features/onboarding';
+import { HandlePracticeOnboarding } from '@/lib/features/onboarding';
 import { AccountsServiceParams as Context } from '../params';
 import { TransactionV1 } from '@/lib/utils';
 
 import {
     transactionDefinition,
-    CreatePracticeEntity,
+    HandlePracticeOnboardingEntity,
     CreateStripeCheckoutSession,
     GetUserDetails,
 } from './transaction';
 
 export function factory(context: Context) {
-    return async function registerProvider(params: CreatePractice.Input) {
+    return async function registerProvider(
+        params: HandlePracticeOnboarding.Input
+    ) {
         return await TransactionV1.executeTransaction(
             transactionDefinition,
             { ...context },
             {
                 getUserDetails: GetUserDetails.factory(params),
-                createPracticeEntity: CreatePracticeEntity.factory(params),
+                createPracticeEntity:
+                    HandlePracticeOnboardingEntity.factory(params),
                 createStripeCheckoutSession:
                     CreateStripeCheckoutSession.factory(params),
             },
