+++ src/lib/services/accounts/service/handle-practice-onboarding/index.ts
@@ -1,0 +1,1 @@
+export * as HandlePracticeOnboarding from './handlePracticeOnboarding';
