+++ src/lib/services/accounts/service/handle-practice-onboarding/transaction/index.ts
@@ -1,4 +1,4 @@
 export { transactionDefinition } from './definition';
 export * as GetUserDetails from './getUserDetails';
-export * as CreatePracticeEntity from './createPracticeEntity';
+export * as HandlePracticeOnboardingEntity from './createPracticeEntity';
 export * as CreateStripeCheckoutSession from './createStripeCheckoutSession';
